Restructure and upload easier to follow examples
