+++ melodies_monet/examples/submit_jobs/run_melodies_monet.py
@@ -5,12 +5,14 @@
 #This is needed to tell matplotlib to use a non-interactive backend and avoid display errors.
 import matplotlib
 matplotlib.use('Agg')
+import sys
+sys.path.append('../../')
 import driver
 import os
 import dask
 an = driver.analysis()
 # -- Update the yaml file below
-an.control = 'control_cmaq-rrfs_surface.yaml'
+an.control = '../yaml/control_cmaq-rrfs_surface-all-short_test_jupyter.yaml'
 an.read_control()
 # -- Lines below make a copy of the namelist in the plot directory for reference later
 cmd = 'cp ' + an.control + ' ' + an.control_dict['analysis']['output_dir']
@@ -21,4 +23,4 @@ an.open_obs()
 an.pair_data()
 # -- Optionally comment out lines below if only want to create plots or stats
 an.plotting()
-#an.stats()
+an.stats()
