+++ melodies_monet/examples/yaml/control_cmaq-rrfs_surface-all-short_test_jupyter.yaml
@@ -76,7 +76,7 @@ model:
 obs:
   airnow: # obs label
     use_airnow: True
-    filename: /scratch2/BMC/rcm1/rhs/MONET/MELODIES-MONET/melodies_monet/data/AIRNOW_20190831_20190831.nc
+    filename: /scratch2/BMC/rcm1/rhs/MONET/MELODIES-MONET/melodies_monet/data/AIRNOW_20190801_20190831.nc
     obs_type: pt_sfc
     variables: #Opt 
       PM2.5:
