+++ melodies_monet/examples/yaml/control_cmaq-rrfs_surface.yaml
@@ -114,7 +114,7 @@ model:
 obs:
   airnow: # obs label
     use_airnow: True
-    filename: data/AIRNOW_20190831_20190831.nc
+    filename: /scratch2/BMC/rcm1/rhs/MONET/MELODIES-MONET/melodies_monet/data/AIRNOW_20190801_20190831.nc
     obs_type: pt_sfc
     variables: #Opt 
       PM2.5:
@@ -265,5 +265,5 @@ stats:
     edges: 'horizontal'
   domain_type: ['all'] #List of domain types: 'all' or any domain in obs file. (e.g., airnow: epa_region, state_name, siteid, etc.) 
   domain_name: ['CONUS'] #List of domain names. If domain_type = all domain_name is used in plot title.
-  data: ['airnow_cmaq_expt','airnow_rrfs_13km','airnow_wrfchem_v4.2'] # make this a list of pairs in obs_model where the obs is the obs label and model is the model_label
+  data: ['airnow_rrfs_r131v1_all','airnow_rrfs_r131v1_wopc','airnow_rrfs_r131v1_pm25','airnow_cmaq_expt','airnow_cmaq_oper'] # make this a list of pairs in obs_model where the obs is the obs label and model is the model_label
