+++ python_env_ymls/cheyenne/cheyenne_monet_env.yml
@@ -1,0 +1,228 @@
+name: monet_py36
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=1_gnu
+  - aggdraw=1.3.12=py36hf6b338f_1
+  - aiobotocore=1.3.1=pyhd8ed1ab_0
+  - aiohttp=3.7.4.post0=py36h8f6f2f9_0
+  - aioitertools=0.7.1=pyhd8ed1ab_0
+  - alsa-lib=1.2.3=h516909a_0
+  - argon2-cffi=20.1.0=py36h8f6f2f9_2
+  - async-timeout=3.0.1=py_1000
+  - attrs=21.2.0=pyhd8ed1ab_0
+  - backports=1.0=py_2
+  - backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
+  - bleach=3.3.0=pyh44b312d_0
+  - bokeh=2.3.2=py36h5fab9bb_0
+  - botocore=1.20.49=pyhd8ed1ab_0
+  - brotlipy=0.7.0=py36h8f6f2f9_1001
+  - bzip2=1.0.8=h7f98852_4
+  - c-ares=1.17.1=h7f98852_1
+  - ca-certificates=2021.5.30=ha878542_0
+  - cartopy=0.19.0.post1=py36hbcbf2fa_0
+  - certifi=2021.5.30=py36h5fab9bb_0
+  - cf_xarray=0.5.2=pyh6c4a22f_0
+  - cffi=1.14.5=py36hc120d54_0
+  - cftime=1.5.0=py36h4d9540e_0
+  - chardet=4.0.0=py36h5fab9bb_1
+  - click=7.1.2=pyh9f0ad1d_0
+  - cloudpickle=1.6.0=py_0
+  - configobj=5.0.6=py_0
+  - contextvars=2.4=py_0
+  - cryptography=3.4.7=py36hb60f036_0
+  - curl=7.77.0=hea6ffbf_0
+  - cycler=0.10.0=py_2
+  - cython=0.29.23=py36hc4f0c31_1
+  - cytoolz=0.11.0=py36h8f6f2f9_3
+  - dask=2021.3.0=pyhd8ed1ab_0
+  - dask-core=2021.3.0=pyhd8ed1ab_0
+  - dbus=1.13.6=h48d8840_2
+  - decorator=5.0.9=pyhd8ed1ab_0
+  - defusedxml=0.7.1=pyhd8ed1ab_0
+  - distributed=2021.3.0=py36h5fab9bb_0
+  - entrypoints=0.3=pyhd8ed1ab_1003
+  - esmf=8.1.1=nompi_hb4f3075_0
+  - esmpy=8.1.1=nompi_py36h7bb7d63_0
+  - expat=2.4.1=h9c3ff4c_0
+  - fontconfig=2.13.1=hba837de_1005
+  - freetype=2.10.4=h0708190_1
+  - fsspec=2021.6.1=pyhd8ed1ab_0
+  - geos=3.9.1=h9c3ff4c_2
+  - gettext=0.19.8.1=h0b5b191_1005
+  - glib=2.68.3=h9c3ff4c_0
+  - glib-tools=2.68.3=h9c3ff4c_0
+  - gst-plugins-base=1.14.0=hbbd80ab_1
+  - gstreamer=1.14.0=h28cd5cc_2
+  - hdf4=4.2.15=h10796ff_3
+  - hdf5=1.10.6=nompi_h6a2412b_1114
+  - heapdict=1.0.1=py_0
+  - icu=58.2=hf484d3e_1000
+  - idna=3.1=pyhd3deb0d_0
+  - idna_ssl=1.1.0=py36h9f0ad1d_1001
+  - immutables=0.15=py36h8f6f2f9_0
+  - importlib-metadata=4.5.0=py36h5fab9bb_0
+  - ipykernel=5.5.5=py36hcb3619a_0
+  - ipython=5.8.0=py36_1
+  - ipython_genutils=0.2.0=py_1
+  - ipywidgets=7.6.3=pyhd3deb0d_0
+  - jbig=2.1=h7f98852_2003
+  - jinja2=3.0.1=pyhd8ed1ab_0
+  - jmespath=0.10.0=pyh9f0ad1d_0
+  - joblib=1.0.1=pyhd8ed1ab_0
+  - jpeg=9d=h36c2ea0_0
+  - jsonschema=3.2.0=pyhd8ed1ab_3
+  - jupyter=1.0.0=py36h5fab9bb_6
+  - jupyter_client=6.1.12=pyhd8ed1ab_0
+  - jupyter_console=5.2.0=py36_1
+  - jupyter_core=4.7.1=py36h5fab9bb_0
+  - jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
+  - kiwisolver=1.3.1=py36h605e78d_1
+  - krb5=1.19.1=hcc1bbae_0
+  - lcms2=2.12=hddcbb42_0
+  - ld_impl_linux-64=2.35.1=hea4e1c9_2
+  - lerc=2.2.1=h9c3ff4c_0
+  - libblas=3.9.0=9_openblas
+  - libcblas=3.9.0=9_openblas
+  - libclang=11.1.0=default_ha53f305_1
+  - libcurl=7.77.0=h2574ce0_0
+  - libdeflate=1.7=h7f98852_5
+  - libedit=3.1.20191231=he28a2e2_2
+  - libev=4.33=h516909a_1
+  - libevent=2.1.10=hcdb4288_3
+  - libffi=3.3=h58526e2_2
+  - libgcc-ng=9.3.0=h2828fa1_19
+  - libgfortran-ng=9.3.0=hff62375_19
+  - libgfortran5=9.3.0=hff62375_19
+  - libglib=2.68.3=h3e27bee_0
+  - libgomp=9.3.0=h2828fa1_19
+  - libiconv=1.16=h516909a_0
+  - liblapack=3.9.0=9_openblas
+  - libllvm11=11.1.0=hf817b99_2
+  - libnetcdf=4.8.0=nompi_hcd642e3_103
+  - libnghttp2=1.43.0=h812cca2_0
+  - libogg=1.3.4=h7f98852_1
+  - libopenblas=0.3.15=pthreads_h8fe5266_1
+  - libopus=1.3.1=h7f98852_1
+  - libpng=1.6.37=h21135ba_2
+  - libpq=13.3=hd57d9b9_0
+  - libsodium=1.0.18=h36c2ea0_1
+  - libssh2=1.9.0=ha56f1ee_6
+  - libstdcxx-ng=9.3.0=h6de172a_19
+  - libtiff=4.3.0=hf544144_1
+  - libuuid=2.32.1=h7f98852_1000
+  - libvorbis=1.3.7=h9c3ff4c_0
+  - libwebp-base=1.2.0=h7f98852_2
+  - libxcb=1.13=h7f98852_1003
+  - libxkbcommon=1.0.3=he3ba5ed_0
+  - libxml2=2.9.12=h03d6c58_0
+  - libzip=1.8.0=h4de3113_0
+  - locket=0.2.0=py_2
+  - lz4-c=1.9.3=h9c3ff4c_0
+  - markupsafe=2.0.1=py36h8f6f2f9_0
+  - matplotlib=3.1.3=py36_0
+  - matplotlib-base=3.1.3=py36hef1b27d_0
+  - mistune=0.8.4=py36h8f6f2f9_1003
+  - monet=v2.2.2=py36h5fab9bb_1
+  - monetio=0.1=pyh9f0ad1d_1
+  - msgpack-python=1.0.2=py36h605e78d_1
+  - multidict=5.1.0=py36h8f6f2f9_1
+  - mysql-common=8.0.25=ha770c72_2
+  - mysql-libs=8.0.25=hfa10184_2
+  - nbconvert=5.6.1=py36h9f0ad1d_1
+  - nbformat=5.1.3=pyhd8ed1ab_0
+  - ncurses=6.2=h58526e2_4
+  - netcdf-fortran=4.5.3=nompi_hf3f1587_104
+  - netcdf4=1.5.7=nompi_py36h3d597d4_100
+  - notebook=6.3.0=py36h5fab9bb_0
+  - nspr=4.30=h9c3ff4c_0
+  - nss=3.67=hb5efdd6_0
+  - numpy=1.19.5=py36h2aa4a07_1
+  - olefile=0.46=pyh9f0ad1d_1
+  - openjpeg=2.4.0=hb52868f_1
+  - openssl=1.1.1k=h7f98852_0
+  - packaging=20.9=pyh44b312d_0
+  - pandas=1.1.5=py36h284efc9_0
+  - pandoc=2.14.0.3=h7f98852_0
+  - pandocfilters=1.4.2=py_1
+  - partd=1.2.0=pyhd8ed1ab_0
+  - patsy=0.5.1=py_0
+  - pcre=8.45=h9c3ff4c_0
+  - pexpect=4.8.0=pyh9f0ad1d_2
+  - pickleshare=0.7.5=py_1003
+  - pillow=8.2.0=py36ha6010c0_1
+  - pip=21.1.2=pyhd8ed1ab_0
+  - proj=7.2.0=h277dcde_2
+  - prometheus_client=0.11.0=pyhd8ed1ab_0
+  - prompt_toolkit=1.0.15=py_1
+  - psutil=5.8.0=py36h8f6f2f9_1
+  - pthread-stubs=0.4=h36c2ea0_1001
+  - ptyprocess=0.7.0=pyhd3deb0d_0
+  - pycparser=2.20=pyh9f0ad1d_2
+  - pydecorate=0.3.1=py_1
+  - pygments=2.9.0=pyhd8ed1ab_0
+  - pykdtree=1.3.4=py36ha112f06_0
+  - pyopenssl=20.0.1=pyhd8ed1ab_0
+  - pyparsing=2.4.7=pyh9f0ad1d_0
+  - pyproj=3.0.1=py36h13f0598_0
+  - pyqt=5.9.2=py36hcca6a23_4
+  - pyqt5-sip=4.19.18=py36hc4f0c31_7
+  - pyresample=1.20.0=py36h84e8a9e_0
+  - pyrsistent=0.17.3=py36h8f6f2f9_2
+  - pyshp=2.1.3=pyh44b312d_0
+  - pysocks=1.7.1=py36h5fab9bb_3
+  - python=3.6.13=hffdb5ce_0_cpython
+  - python-dateutil=2.8.1=py_0
+  - python-stratify=0.1.1=py36h4d9540e_1004
+  - python_abi=3.6=2_cp36m
+  - pytz=2021.1=pyhd8ed1ab_0
+  - pyyaml=5.4.1=py36h8f6f2f9_0
+  - pyzmq=22.1.0=py36h7068817_0
+  - qt=5.9.7=h5867ecd_1
+  - qtconsole=5.1.0=pyhd8ed1ab_0
+  - qtpy=1.9.0=py_0
+  - readline=8.1=h46c0cb4_0
+  - s3fs=2021.6.1=pyhd8ed1ab_0
+  - scipy=1.5.3=py36h9e8f40b_0
+  - seaborn=0.11.1=hd8ed1ab_1
+  - seaborn-base=0.11.1=pyhd8ed1ab_1
+  - send2trash=1.7.1=pyhd8ed1ab_0
+  - setuptools=49.6.0=py36h5fab9bb_3
+  - shapely=1.7.1=py36h3bbdef7_5
+  - simplegeneric=0.8.1=py_1
+  - sip=4.19.8=py36hf484d3e_1000
+  - six=1.16.0=pyh6c4a22f_0
+  - sortedcontainers=2.4.0=pyhd8ed1ab_0
+  - sqlite=3.36.0=h9cd32fc_0
+  - statsmodels=0.12.2=py36h92226af_0
+  - tblib=1.7.0=pyhd8ed1ab_0
+  - terminado=0.10.1=py36h5fab9bb_0
+  - testpath=0.5.0=pyhd8ed1ab_0
+  - tk=8.6.10=h21135ba_1
+  - toolz=0.11.1=py_0
+  - tornado=6.1=py36h8f6f2f9_1
+  - traitlets=4.3.3=py36h9f0ad1d_1
+  - typing-extensions=3.10.0.0=hd8ed1ab_0
+  - typing_extensions=3.10.0.0=pyha770c72_0
+  - urllib3=1.26.6=pyhd8ed1ab_0
+  - wcwidth=0.2.5=pyh9f0ad1d_2
+  - webencodings=0.5.1=py_1
+  - wheel=0.36.2=pyhd3deb0d_0
+  - widgetsnbextension=3.5.1=py36h5fab9bb_4
+  - wrapt=1.12.1=py36h8f6f2f9_3
+  - wrf-python=1.3.2.5=py36h2663098_0
+  - xarray=0.18.2=pyhd8ed1ab_0
+  - xesmf=0.5.3=pyhd8ed1ab_0
+  - xorg-libxau=1.0.9=h7f98852_0
+  - xorg-libxdmcp=1.1.3=h7f98852_0
+  - xz=5.2.5=h516909a_1
+  - yaml=0.2.5=h516909a_0
+  - yarl=1.6.3=py36h8f6f2f9_1
+  - zeromq=4.3.4=h9c3ff4c_0
+  - zict=2.0.0=py_0
+  - zipp=3.4.1=pyhd8ed1ab_0
+  - zlib=1.2.11=h516909a_1010
+  - zstd=1.5.0=ha95c52a_0
+prefix: /glade/u/home/buchholz/miniconda3/envs/monet_py36
