--- python_env_ymls/environment.yml
@@ -1,214 +1,0 @@
-name: py36_monet_def_test
-channels:
-  - defaults
-  - https://anaconda.rdhpcs.noaa.gov/pkgs/main/
-  - https://anaconda.rdhpcs.noaa.gov/pkgs/free/
-  - https://anaconda.rdhpcs.noaa.gov/pkgs/r/
-  - https://anaconda.rdhpcs.noaa.gov/conda-forge/
-  - https://anaconda.rdhpcs.noaa.gov/ska/
-  - https://anaconda.rdhpcs.noaa.gov/cdat/
-  - https://anaconda.rdhpcs.noaa.gov/ncar/
-  - https://anaconda.rdhpcs.noaa.gov/pytorch/
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=1_llvm
-  - aggdraw=1.3.12=py36hf6b338f_1
-  - argon2-cffi=20.1.0=py36h7b6447c_1
-  - async_generator=1.10=py_0
-  - attrs=20.3.0=pyhd3eb1b0_0
-  - backcall=0.2.0=py_0
-  - blas=2.106=openblas
-  - bleach=3.2.1=py_0
-  - bokeh=2.3.1=py36h5fab9bb_0
-  - boto3=1.16.42=pyhd3eb1b0_0
-  - botocore=1.19.43=pyhd3eb1b0_0
-  - brotlipy=0.7.0=py36h27cfd23_1003
-  - bzip2=1.0.8=h7b6447c_0
-  - ca-certificates=2020.12.8=h06a4308_0
-  - cartopy=0.19.0=py36hbcbf2fa_0
-  - certifi=2020.12.5=py36h06a4308_0
-  - cf_xarray=0.5.1=pyh44b312d_0
-  - cffi=1.14.4=py36h261ae71_0
-  - cftime=1.3.0=py36h6323ea4_0
-  - click=7.1.2=pyh9f0ad1d_0
-  - cloudpickle=1.6.0=py_0
-  - configobj=5.0.6=py_0
-  - contextvars=2.4=py_0
-  - cryptography=3.3.1=py36h3c74f83_0
-  - curl=7.71.1=hbc83047_1
-  - cycler=0.10.0=py_2
-  - cython=0.29.21=py36h2531618_0
-  - cytoolz=0.11.0=py36h8f6f2f9_3
-  - dask=2021.3.0=pyhd8ed1ab_0
-  - dask-core=2021.3.0=pyhd8ed1ab_0
-  - dbus=1.13.18=hb2f20db_0
-  - decorator=4.4.2=py_0
-  - defusedxml=0.6.0=py_0
-  - distributed=2021.3.0=py36h5fab9bb_0
-  - docutils=0.15.2=py36_0
-  - entrypoints=0.3=py36_0
-  - esmf=8.1.0=mpi_mpich_h224d73f_100
-  - esmpy=8.1.0=mpi_mpich_py36hcd78dbd_100
-  - expat=2.2.10=he6710b0_2
-  - fontconfig=2.13.0=h9420a91_0
-  - freetype=2.10.4=h0708190_1
-  - fsspec=2021.4.0=pyhd8ed1ab_0
-  - geos=3.9.1=h9c3ff4c_2
-  - glib=2.66.1=h92f7085_0
-  - gst-plugins-base=1.14.0=hbbd80ab_1
-  - gstreamer=1.14.0=hb31296c_0
-  - hdf4=4.2.13=h3ca952b_2
-  - hdf5=1.10.6=mpi_mpich_h996c276_1014
-  - heapdict=1.0.1=py_0
-  - icu=58.2=he6710b0_3
-  - idna=2.10=py_0
-  - immutables=0.15=py36h8f6f2f9_0
-  - importlib-metadata=2.0.0=py_1
-  - importlib_metadata=2.0.0=1
-  - ipykernel=5.3.4=py36h5ca1d4c_0
-  - ipython=7.16.1=py36h5ca1d4c_0
-  - ipython_genutils=0.2.0=pyhd3eb1b0_1
-  - ipywidgets=7.6.0=pyhd3eb1b0_1
-  - jedi=0.18.0=py36h06a4308_0
-  - jinja2=2.11.3=pyh44b312d_0
-  - jmespath=0.10.0=py_0
-  - joblib=1.0.0=pyhd3eb1b0_0
-  - jpeg=9d=h36c2ea0_0
-  - jsonschema=3.2.0=py_2
-  - jupyter=1.0.0=py36_7
-  - jupyter_client=6.1.7=py_0
-  - jupyter_console=6.2.0=py_0
-  - jupyter_core=4.7.0=py36h06a4308_0
-  - jupyterlab_pygments=0.1.2=py_0
-  - kiwisolver=1.3.1=py36h605e78d_1
-  - krb5=1.18.2=h173b8e3_0
-  - ld_impl_linux-64=2.35.1=hea4e1c9_2
-  - libblas=3.9.0=6_openblas
-  - libcblas=3.9.0=6_openblas
-  - libcurl=7.71.1=h20c2e04_1
-  - libedit=3.1.20191231=h14c3975_1
-  - libffi=3.3=he6710b0_2
-  - libgcc-ng=9.3.0=h2828fa1_19
-  - libgfortran-ng=9.3.0=hff62375_19
-  - libgfortran5=9.3.0=hff62375_19
-  - libgomp=9.3.0=h2828fa1_19
-  - liblapack=3.9.0=6_openblas
-  - liblapacke=3.9.0=6_openblas
-  - libnetcdf=4.7.4=mpi_mpich_hdef422e_7
-  - libopenblas=0.3.12=pthreads_h4812303_1
-  - libpng=1.6.37=h21135ba_2
-  - libsodium=1.0.18=h7b6447c_0
-  - libssh2=1.9.0=h1ba5d50_1
-  - libstdcxx-ng=9.3.0=h6de172a_19
-  - libtiff=4.1.0=h2733197_1
-  - libuuid=1.0.3=h1bed415_2
-  - libxcb=1.14=h7b6447c_0
-  - libxml2=2.9.10=hb55368b_3
-  - llvm-openmp=11.1.0=h4bd325d_1
-  - locket=0.2.0=py_2
-  - lz4-c=1.9.2=he1b5a44_3
-  - markupsafe=1.1.1=py36h8f6f2f9_3
-  - matplotlib=3.1.3=py36_0
-  - matplotlib-base=3.1.3=py36hef1b27d_0
-  - mistune=0.8.4=py36h7b6447c_0
-  - monet=v2.2.2=py36h5fab9bb_1
-  - monetio=0.1=pyh9f0ad1d_1
-  - mpi=1.0=mpich
-  - mpi4py=3.0.3=py36h7b8b12a_7
-  - mpich=3.4.1=h846660c_104
-  - msgpack-python=1.0.2=py36h605e78d_1
-  - nbclient=0.5.1=py_0
-  - nbconvert=6.0.7=py36_0
-  - nbformat=5.0.8=py_0
-  - ncurses=6.2=he6710b0_1
-  - nest-asyncio=1.4.3=pyhd3eb1b0_0
-  - netcdf-fortran=4.5.3=mpi_mpich_hafa3f36_3
-  - netcdf4=1.5.6=nompi_py36hb848c25_102
-  - notebook=6.1.6=py36h06a4308_0
-  - numpy=1.19.2=py36h6163131_0
-  - numpy-base=1.19.2=py36h75fe3a5_0
-  - olefile=0.46=pyh9f0ad1d_1
-  - openssl=1.1.1i=h27cfd23_0
-  - packaging=20.9=pyh44b312d_0
-  - pandas=1.1.5=py36h284efc9_0
-  - pandoc=2.11=hb0f4dca_0
-  - pandocfilters=1.4.3=py36h06a4308_1
-  - parso=0.7.0=py_0
-  - partd=1.2.0=pyhd8ed1ab_0
-  - patsy=0.5.1=py_0
-  - pcre=8.44=he6710b0_0
-  - pexpect=4.8.0=pyhd3eb1b0_3
-  - pickleshare=0.7.5=pyhd3eb1b0_1003
-  - pillow=7.1.2=py36hb39fc2d_0
-  - pip=20.3.3=py36h06a4308_0
-  - proj=7.2.0=h8b9fe22_0
-  - prometheus_client=0.9.0=pyhd3eb1b0_0
-  - prompt-toolkit=3.0.8=py_0
-  - prompt_toolkit=3.0.8=0
-  - psutil=5.8.0=py36h8f6f2f9_1
-  - ptyprocess=0.6.0=pyhd3eb1b0_2
-  - pycparser=2.20=py_2
-  - pydecorate=0.3.1=py_1
-  - pygments=2.7.3=pyhd3eb1b0_0
-  - pykdtree=1.3.4=py36ha112f06_0
-  - pyopenssl=20.0.1=pyhd3eb1b0_1
-  - pyparsing=2.4.7=pyh9f0ad1d_0
-  - pyproj=3.0.1=py36h13f0598_0
-  - pyqt=5.9.2=py36h05f1152_2
-  - pyresample=1.19.0=py36h84e8a9e_0
-  - pyrsistent=0.17.3=py36h7b6447c_0
-  - pyshp=2.1.3=pyh44b312d_0
-  - pysocks=1.7.1=py36h06a4308_0
-  - python=3.6.12=hcff3b4d_2
-  - python-dateutil=2.8.1=py_0
-  - python-stratify=0.1.1=py36h92226af_1003
-  - python_abi=3.6=1_cp36m
-  - pytz=2021.1=pyhd8ed1ab_0
-  - pyyaml=5.4.1=py36h8f6f2f9_0
-  - pyzmq=20.0.0=py36h2531618_1
-  - qt=5.9.7=h5867ecd_1
-  - qtconsole=4.7.7=py_0
-  - qtpy=1.9.0=py_0
-  - readline=8.0=h7b6447c_0
-  - s3fs=0.4.2=py_0
-  - s3transfer=0.3.3=pyhd3eb1b0_3
-  - scipy=1.5.3=py36h9e8f40b_0
-  - seaborn=0.11.1=hd8ed1ab_1
-  - seaborn-base=0.11.1=pyhd8ed1ab_1
-  - send2trash=1.5.0=pyhd3eb1b0_1
-  - setuptools=51.0.0=py36h06a4308_2
-  - shapely=1.7.1=py36h93b233e_4
-  - sip=4.19.8=py36hf484d3e_0
-  - six=1.15.0=pyh9f0ad1d_0
-  - sortedcontainers=2.3.0=pyhd8ed1ab_0
-  - sqlite=3.33.0=h62c20be_0
-  - statsmodels=0.12.2=py36h92226af_0
-  - tblib=1.7.0=pyhd8ed1ab_0
-  - terminado=0.9.1=py36_0
-  - testpath=0.4.4=py_0
-  - tk=8.6.10=hbc83047_0
-  - toolz=0.11.1=py_0
-  - tornado=6.1=py36h8f6f2f9_1
-  - traitlets=4.3.3=py36_0
-  - typing_extensions=3.7.4.3=py_0
-  - urllib3=1.25.11=py_0
-  - wcwidth=0.2.5=py_0
-  - webencodings=0.5.1=py36_1
-  - wheel=0.36.2=pyhd3eb1b0_0
-  - widgetsnbextension=3.5.1=py36_0
-  - wrapt=1.12.1=py36h8f6f2f9_3
-  - wrf-python=1.3.2.5=py36h2663098_0
-  - xarray=0.17.0=pyhd8ed1ab_0
-  - xesmf=0.5.3=pyhd8ed1ab_0
-  - xz=5.2.5=h7b6447c_0
-  - yaml=0.2.5=h516909a_0
-  - zeromq=4.3.3=he6710b0_3
-  - zict=2.0.0=py_0
-  - zipp=3.4.0=pyhd3eb1b0_0
-  - zlib=1.2.11=h7b6447c_3
-  - zstd=1.4.5=h6597ccf_2
-  - pip:
-    - msgpack==1.0.2
-    - stratify==0.1.1
-prefix: /scratch2/BMC/rcm1/rhs/anaconda3/envs/py36_monet_def_test
-
